Remove all interface App type
